Refactor cabinet script into modular structure
